Changed LastPrice to float

+++ Asset Management Platform/SecurityClasses/CorporateBond.cs
@@ -8,13 +8,13 @@ namespace Asset_Management_Platform.SecurityClasses
 {
     class CorporateBond : FixedIncome
     {
-        public CorporateBond(string cusip, string ticker, string description, double lastPrice, double yield) 
+        public CorporateBond(string cusip, string ticker, string description, float lastPrice, double yield) 
             : base(cusip, ticker, description, lastPrice, yield)
         {
             SecurityType = "Corporate Bond";
         }
 
-        public CorporateBond(string cusip, string ticker, string description, double lastPrice, double yield, double coupon, string issuer, string rating)
+        public CorporateBond(string cusip, string ticker, string description, float lastPrice, double yield, double coupon, string issuer, string rating)
             : base(cusip, ticker, description, lastPrice, yield, coupon, issuer, rating)
         {
             SecurityType = "Corporate Bond";
